+++ person.rb
@@ -1,8 +1,4 @@
-class Nameable
-  def correct_name
-    raise NotImplementedError
-  end
-end
+require './nameable'
 
 class Person < Nameable
   attr_accessor :name, :age
@@ -28,28 +24,5 @@ class Person < Nameable
 
   def of_age?
     @age >= 18
-  end
-end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.length > 10 ? @nameable.correct_name[0...10] : @nameable.correct_name
   end
 end
